// src/types/ApiTypes.ts
export interface IRequestEvents {
    type: 'company_calendar',
    ownerId: '',
    from: dayjs.Dayjs,
    to: dayjs.Dayjs,
};